% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-harmonicity.R
\name{pc_harmonicity}
\alias{pc_harmonicity}
\alias{pc_harmonicity.default}
\alias{pc_harmonicity.pc_set}
\alias{pc_harmonicity.pc_milne_spectrum}
\title{Pitch-class harmonicity}
\usage{
pc_harmonicity(x, method = "kl", num_harmonics = 12, rho = 0.75,
  sigma = 6.83, ...)

\method{pc_harmonicity}{default}(x, method = "kl", num_harmonics = 12,
  rho = 0.75, sigma = 6.83, array_dim = 1200, ...)

\method{pc_harmonicity}{pc_set}(x, method = "kl", num_harmonics = 12,
  rho = 0.75, sigma = 6.83, array_dim = 1200, ...)

\method{pc_harmonicity}{pc_milne_spectrum}(x, method = "kl",
  num_harmonics = 12, rho = 0.75, sigma = 6.83, ...)
}
\arguments{
\item{x}{Object to analyse.}

\item{method}{(Character scalar) Method to use.
\itemize{
\item \code{"kl"} (default) delivers the Kullback-Leibler method of
\insertCite{Harrison2018;textual}{har18}.
\item \code{"peak"} delivers the peak-value method of
\insertCite{Milne2013;textual}{har18}.
}}

\item{num_harmonics}{(Integerish scalar)
Number of harmonics to use when expanding tones into their implied harmonics,
and when defining the harmonic template
(including the fundamental frequency).
Defaults to 12, after
\insertCite{Milne2016;textual}{har18}.}

\item{rho}{(Numeric scalar)
Roll-off parameter for harmonic expansion.
Defaults to 0.75, after
\insertCite{Milne2016;textual}{har18}.}

\item{sigma}{(Numeric scalar)
Standard deviation of the Gaussian smoothing distribution (cents).
Defaults to 6.83, after
\insertCite{Milne2016;textual}{har18}.}

\item{...}{Arguments passed to specific methods.}

\item{array_dim}{(Integerish scalar)
Dimensionality of the pitch-class spectrum array.
Defaults to 1200, after
\insertCite{Milne2016;textual}{har18}.}
}
\value{
Pitch-class harmonicity, as a numeric scalar.
}
\description{
Gets the pitch-class harmonicity of an input sonority, after
\insertCite{Harrison2018;textual}{har18} and
\insertCite{Milne2013;textual}{har18}.
}
\note{
This algorithm makes use of \code{\link[hrep]{pc_milne_spectrum}()}
as defined in the \code{hrep} package.
}
\examples{
pc_harmonicity(c(0, 4, 7))
pc_harmonicity(c(0, 3, 7))
pc_harmonicity(c(0, 3, 6))
}
\references{
\insertAllCited{}
}
