% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{peak}
\alias{peak}
\alias{peak.pc_milne_spectrum}
\title{Peak}
\usage{
peak(x)

\method{peak}{pc_milne_spectrum}(x)
}
\arguments{
\item{x}{Object to analyse.}
}
\value{
The object's peak value, as a numeric scalar.
}
\description{
Gets the peak value of an object.
}
